project/taskrunner/requirements.txt
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
google-api-python-client==1.10.0
google-auth==1.19.2
google-cloud-bigquery==1.15.0
google-cloud-storage==1.16.1
oauth2client==4.1.3
psycopg2==2.8.5
PyYAML==5.3.1
requests==2.24.0
redis>=3
rq>=0.13
sentry_sdk==0.17.8
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
statsd==3.3.0
